+++ packages/react-docs/CHANGELOG.md
@@ -5,3 +5,6 @@
     * Move example out into a separate sub-package
     * Consolidate all `console.log` calls into `logUtils` in the `@0xproject/utils` package (#452)
     * Handle `reflection` type rendering so that anonymous function type declarations render properly
+    * Rename `MethodSignature` to `Signature` and change it's props so that it can be used to render method and function signatures.
+    * Rename `MethodBlock` to `SignatureBlock` since it is not used to render method and function signature blocks.
+    * Add support for documenting exported functions.